darkpoolv2: settlement: Use separate libraries for privacy rings (#148)
### Purpose
This PR refactors the settlement lib to branch earlier in the settlement verification process on a per-ring basis. That is, we will define privacy-ring-specific settlement verification libraries and branch at the top level into those.

This PR implements the beginning of the ring 0 library, bringing over logic that was present in `SettlementLib.sol` before.

### Testing
- [x] All test pass