darkpoolv2: types: Transfers: Implement the settlement transfer list (#151)
### Purpose
This PR changes the approach for executing ERC transfers. In specific, we allocate a dynamic-like list of deposits and withdrawals up-front and allow the validation logic to register transfers into this list as they progress through their ring-specific logic.

The darkpool then executes all transfers at once at the end of the settlement method.

A similar approach will be taken for proofs: let handlers register them in a semi-dynamic list and batch verify them all at the end.

### Todo
- Define state update methods for the handlers which register transfers here
- Test full match path with balance differences

### Testing
- [x] Unit tests pass