darkpoolv2: VerificationList: Implement `VerificationList` type (#159)
### Purpose
This PR creates a new `VerificationList` type which is analogous to the `SettlementTransfers` type created before. That is, this type allows verification handlers to dynamically register proofs during their execution. These proofs will be batch verified at the end of the transaction.

I then generalized the `SettlementTransfers` allocation into a `SettlementContext` allocation, which tracks both transfers and proofs.

### Todo
- Verify proofs in the verification list
- Register proofs in the native settled private intent handlers

### Testing
- [x] Existing unit tests pass