darkpoolv2: settlement: Ring 2 intent authorization first fill (#169)
### Purpose
This PR adds intent authorization logic to the darkpool for Ring 2 users. On the first fill, we verify that the user has spent a one-time key to sign the commitment to the intent as well as the updated one-time key.

For subsequent fills, we need only verify a validity proof.

### Testing
- [x] Unit tests pass