darkpoolv2: obligation: Move obligation bundle to top level argument (#173)
### Purpose
This PR updates the obligation bundle to store _both_ the obligations in a trade, rather than encoding the obligation bundle _within_ a settlement bundle. We will use this pattern to encode a settlement proof in the case that the obligation is private. 

This will be used in place of bundle-specific settlement proofs currently in use.

### Testing
- [x] All unit tests pass